requirements.txt
langgraph==0.0.15
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.3,<2.6.0
langchain-core>=0.1.27
langchain>=0.1.9
langchain-openai>=0.0.8
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.18.1
webdriver-manager>=4.0.1
duckduckgo-search>=4.1.1
trafilatura>=1.6.3 